backend/requirements.txt
django<5.0
djangorestframework
django-cors-headers
reportlab
pyhanko[crypto]
asn1crypto
cryptography
psycopg2-binary
gunicorn
whitenoise
